TechDebt moved tests to subfolder as well.
